requirements.txt
opencv-python>=4.5.0
python-dotenv>=0.19.0
cloudinary>=1.30.0
